requirements.txt
backcall==0.1.0
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
Fiona==1.7.13
Flask==0.12.2
Flask-API==1.0
h5py==2.8.0
httplib2==0.12.0
jedi==0.13.1
jsonschema==3.0.0a3
MarkupSafe==1.0
matplotlib==2.2.3
more-itertools==4.3.0
numpy==1.15.4
oauth2client==4.1.3
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
psycopg2==2.7.6.1
ptyprocess==0.6.0
PuLP==1.6.9
Pygments==2.3.0
pyparsing==2.3.0
pystan==2.17.1.0
python-dateutil==2.7.5
python-dotenv==0.10.0
pytz==2018.7
pyzmq==17.1.2
opencv-python==3.4.4.19
Rtree==0.8.3
scikit-learn==0.20.1
scipy==1.1.0
tensorboard==1.12.0
tensorflow==1.12.0
webencodings==0.5.1
Werkzeug==0.12.2
gevent==1.3.7
gunicorn==19.9.0
